legacy: the names, formats and arithmetic a GCC build objects to
NULL without stddef.h in two files; a static named accept, which is Winsock's
the moment winsock2.h is in scope; locals named near and exit; a DWORD printed
with %u; a PDH_STATUS compared to an unsigned constant; a tick subtraction
done in signed arithmetic, which is undefined on the wrap it was written for;
and enhanced_input relying on a pragma to link xinput where controller_input
does both.

Compiled and tested.

+++ legacy/src/mods/diagnostics/diag_log.c
@@ -37,8 +37,9 @@ static void write_raw(const char *line)
     DWORD elapsed = timeGetTime() - diag_state.started_at;
 
     if (diag_state.file != NULL) {
-        fprintf(diag_state.file, "[%4u.%03u] %s\n",
-                elapsed / MILLISECONDS_PER_SECOND, elapsed % MILLISECONDS_PER_SECOND, line);
+        fprintf(diag_state.file, "[%4lu.%03lu] %s\n",
+                (unsigned long)(elapsed / MILLISECONDS_PER_SECOND),
+                (unsigned long)(elapsed % MILLISECONDS_PER_SECOND), line);
         fflush(diag_state.file);
     }
     if (diag_state.also_to_main_log) {
